feat (error handlering): add profile errors, user errors and shipping price errors

+++ src/domain/service/category.service.ts
@@ -3,6 +3,9 @@ import { IRepository } from 'src/domain/outgoing/repository.interface';
 import { ICategoryService } from 'src/domain/incoming/category.service.interface';
 import { PaginatedResult } from 'src/domain/model/paginated-result';
 import { Category } from 'src/domain/model/category/category';
+import { CategoryNotFoundError, DuplicateCategoryError } from '../error/category-errors';
+import { DomainError } from '../error/domain-error';
+import { ResponseCode } from '../error/response-code.enum';
 
 /**
  * Category Service
@@ -37,34 +40,48 @@ export class CategoryService implements ICategoryService<Category> {
 
   // Get a single category
   async getById(id: string): Promise<Category> {
-    const category: Category = await this.categoryRepository.getById(id);
-    return category;
+    const entity: Category = await this.categoryRepository.getById(id);
+    if (!entity || entity === null) throw new CategoryNotFoundError();
+    return entity;
   };
 
   async create(entity: Category): Promise<Category> {
+    try{
     const entityNew: Promise<Category> = this.categoryRepository.create(entity);
     return entityNew;
+  } catch (error) {
+    if (error.code && error.code === 11000) {
+      throw new DuplicateCategoryError(`Database error: Duplicate key error collection or index problem. ${error.message}`);
+    }
+    throw new DomainError(ResponseCode.INTERNAL_SERVER_ERROR, error.message, '', error); 
+  }
   };
 
   // Delete category return this.labelModel.deleteOne({ osCode }).exec();
   async delete(id: string): Promise<boolean> {
+    const found: boolean = await this.categoryRepository.hasById(id);
+    if (!found) throw new CategoryNotFoundError();
     const deleted: boolean = await this.categoryRepository.delete(id);
     return deleted;
   };
 
   // Put a single category
   async updateById(id: string, category: Category): Promise<boolean> {
+    const found: boolean = await this.categoryRepository.hasById(id);
+    if (!found) throw new CategoryNotFoundError();
     const updatedProduct: boolean = await this.categoryRepository.updateById(id, category);
     return updatedProduct;
   };
 
   async getByQuery(query: any): Promise<Category> {
-    const cat =  await this.categoryRepository.getByQuery(query);
-    return cat;
+    const entity: Category =  await this.categoryRepository.getByQuery(query);
+    if (!entity || entity === null) throw new CategoryNotFoundError();
+    return entity;
   };
 
   async update(query: any, valuesToSet: any): Promise<boolean> {
     const updatedProduct: boolean = await this.categoryRepository.update(query, valuesToSet);
+    if (!updatedProduct) throw new CategoryNotFoundError();
     return updatedProduct;
   };
